+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed a typo in the parser.
 - Switched to output window instead of terminal for the Compile SM command.
+- Fixed malformed code outline (see #179).
 
 ## [3.0.5]